requirements_windows.txt
# =============================================================================
# 🚀 REQUIREMENTS PARA WINDOWS - TRADING BOT CON TCN
# =============================================================================
# Dependencias optimizadas para Windows con TensorFlow 2.15

# === MACHINE LEARNING Y DEEP LEARNING ===
tensorflow==2.15.0              # TensorFlow 2.15 para Windows
keras==2.15.0                   # Keras compatible con TF 2.15
numpy==1.24.3                   # NumPy compatible con TF 2.15
pandas==2.1.1                   # Análisis de datos
scikit-learn==1.3.1             # ML utilities

# === ANÁLISIS TÉCNICO ===
ta==0.10.2                      # Technical Analysis Library
TA-Lib==0.4.28                  # TA-Lib (requiere instalación especial en Windows)
pandas-ta==0.3.14b0             # Pandas Technical Analysis

# === BINANCE Y CRYPTO APIs ===
python-binance==1.0.19          # Cliente oficial Binance
ccxt==4.1.77                    # Unified crypto exchange library
websockets==12.0                # WebSocket client para streams

# === API Y WEB REQUESTS ===
requests==2.31.0                # HTTP requests
aiohttp==3.9.1                  # Async HTTP client/server
httpx==0.25.2                   # Modern HTTP client
fastapi==0.104.1                # Fast API framework (opcional)
uvicorn==0.24.0                 # ASGI server (opcional)

# === BASE DE DATOS ===
SQLAlchemy==2.0.23              # ORM para base de datos
aiosqlite==0.19.0               # SQLite async support
psycopg2-binary==2.9.9          # PostgreSQL (opcional)

# === CONFIGURACIÓN Y ENTORNO ===
python-dotenv==1.0.0            # Variables de entorno
pydantic==2.5.1                 # Data validation
pydantic-settings==2.1.0        # Settings management

# === NOTIFICACIONES ===
discord-webhook==1.3.0          # Discord notifications
requests-oauthlib==1.3.1        # OAuth para APIs

# === UTILIDADES DE TIEMPO ===
python-dateutil==2.8.2          # Date utilities
pytz==2023.3                    # Timezone support
croniter==1.4.1                 # Cron expressions

# === LOGGING Y MONITOREO ===
structlog==23.2.0               # Structured logging
colorama==0.4.6                 # Colored terminal output (Windows)
rich==13.7.0                    # Rich text and beautiful formatting

# === DESARROLLO Y TESTING ===
pytest==7.4.3                   # Testing framework
pytest-asyncio==0.21.1          # Async testing
pytest-cov==4.1.0               # Test coverage
black==23.11.0                  # Code formatting
isort==5.12.0                   # Import sorting

# === CRIPTOGRAFÍA Y SEGURIDAD ===
cryptography==41.0.7            # Cryptographic recipes
bcrypt==4.1.1                   # Password hashing
PyJWT==2.8.0                    # JSON Web Tokens

# === UTILIDADES GENERALES ===
click==8.1.7                    # CLI creation toolkit
tqdm==4.66.1                    # Progress bars
schedule==1.2.0                 # Job scheduling
pathlib                         # Path utilities (built-in Python 3.4+)

# === PROCESAMIENTO DE DATOS ===
scipy==1.11.4                   # Scientific computing
joblib==1.3.2                   # Efficient serialization

# === VISUALIZACIÓN (OPCIONAL) ===
matplotlib==3.8.2               # Plotting library
plotly==5.17.0                  # Interactive plots
seaborn==0.13.0                 # Statistical visualization

# === PERFORMANCE ===
numba==0.58.1                   # JIT compiler para NumPy
cython==3.0.6                   # C extensions for Python

# === HERRAMIENTAS WINDOWS ESPECÍFICAS ===
pywin32==306; sys_platform == "win32"    # Windows APIs
colorama==0.4.6                 # Cross-platform colored terminal

# === ASYNC Y CONCURRENCIA ===
asyncio                         # Async I/O (built-in Python 3.4+)
aiofiles==23.2.1                # Async file operations
aiodns==3.1.1                   # Async DNS resolution

# === JSON Y SERIALIZACIÓN ===
ujson==5.8.0                    # Ultra-fast JSON parser
orjson==3.9.10                  # Fast JSON library

# === NETWORKING ===
urllib3==2.0.7                  # HTTP library
certifi==2023.11.17             # SSL certificates

# =============================================================================
# 📝 NOTAS DE INSTALACIÓN PARA WINDOWS
# =============================================================================

# Para instalar TA-Lib en Windows:
# 1. Opción 1 - Usar wheel precompilado:
#    pip install https://github.com/cgohlke/talib-build/releases/download/v0.4.28/TA_Lib-0.4.28-cp310-cp310-win_amd64.whl
#
# 2. Opción 2 - Si tienes Visual Studio:
#    - Instalar Visual Studio Build Tools
#    - pip install TA-Lib
#
# 3. Opción 3 - Usar conda:
#    conda install -c conda-forge ta-lib

# Para verificar instalación de TensorFlow:
# python -c "import tensorflow as tf; print('TensorFlow version:', tf.__version__); print('GPU available:', tf.config.list_physical_devices('GPU'))"

# Comando de instalación completa:
# pip install --upgrade pip setuptools wheel
# pip install -r requirements_windows.txt

# =============================================================================
# 🚨 DEPENDENCIAS CRÍTICAS PARA EL TRADING BOT
# =============================================================================

# Estas son las dependencias ESENCIALES que deben instalarse:
# 1. tensorflow==2.15.0      - Para modelos TCN
# 2. python-binance==1.0.19  - Para API de Binance  
# 3. pandas==2.1.1           - Para procesamiento de datos
# 4. numpy==1.24.3           - Para cálculos numéricos
# 5. SQLAlchemy==2.0.23      - Para base de datos
# 6. python-dotenv==1.0.0    - Para configuración
# 7. requests==2.31.0        - Para HTTP requests
# 8. ta==0.10.2              - Para indicadores técnicos 